refactor(work): split oversized modules into focused submodules (GH-610)
Break down tdd-phase-state.js (1515 lines), implement-gate.js (1142 lines),
and enforce-tdd-on-stop.js into focused submodules under new directories.
Original files now serve as thin public-surface wrappers re-exporting from
submodules. Remove 3 entries from .quality-exceptions for split files that
now pass static-quality max-lines gate. 281 tests pass, quality gate clean.
